requirements.txt
# PicGate - AI Image Gateway for OpenWebUI

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
jinja2>=3.1.0
python-multipart>=0.0.6
httpx>=0.25.0
boto3>=1.33.0
bcrypt>=4.1.0
python-dotenv>=1.0.0
